src/main/java/com/rachel/account/service/AccountService.java
package com.rachel.account.service;

import com.rachel.account.model.Account;
import com.rachel.account.model.User;
import com.rachel.account.repository.AccountRepository;
import com.rachel.account.repository.UserRepository;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.Optional;

@Service
@Transactional(readOnly = true)
public class AccountService {

    private final AccountRepository accountRepository;

    public AccountService(AccountRepository accountRepository) {
        this.accountRepository = accountRepository;
    }

    public Optional<Account> getAccount(Long accountId) {
        return accountRepository.findById(accountId);
    }

}
